+++ backend/src/main/java/com/ttp/learning_web/learningPlatform/repository/ProgressRepository.java
@@ -6,16 +6,16 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProgressRepository extends JpaRepository<Progress, Integer> {
-    void deleteByProgressId(Integer progressId);
+public interface ProgressRepository extends JpaRepository<Progress, Long> {
+    void deleteByProgressId(Long progressId);
 
-    Optional<Progress> findByProgressId(Integer progressId);
+    Optional<Progress> findByProgressId(Long progressId);
 
-    List<Progress> findByCourse_CourseId(Integer courseId);
+    List<Progress> findByCourse_CourseId(Long courseId);
 
-    List<Progress> findByUser_UserId(Integer userId);
+    List<Progress> findByUser_UserId(Long userId);
 
-    List<Progress> findByCourse_CourseIdAndUser_UserId(Integer courseId, Integer userId);
+    List<Progress> findByCourse_CourseIdAndUser_UserId(Long courseId, Long userId);
 
-    Optional<Progress> findByCourse_CourseIdAndUser_UserIdAndSkill_SkillId(Integer courseId, Integer userId, Integer bubbleId);
+    Optional<Progress> findByCourse_CourseIdAndUser_UserIdAndSkill_SkillId(Long courseId, Long userId, Long bubbleId);
 }
